+++ focussight/__init__.py
@@ -2,5 +2,12 @@
 
 from .tracker import run_focus_tracker
 from .summary import summarize_file, summarize_directory
+from .ops_report import build_ops_report, render_ops_report
 
-__all__ = ["run_focus_tracker", "summarize_file", "summarize_directory"]
+__all__ = [
+	"run_focus_tracker",
+	"summarize_file",
+	"summarize_directory",
+	"build_ops_report",
+	"render_ops_report",
+]